rename files and update gitignore
